pyproject.toml
[project]
name = "scripta-lambda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "httpx>=0.27.2",
    "json-log-formatter>=1.1",
    "langchain-community>=0.3.4",
    "langchain-core>=0.3.15",
    "langchain-groq>=0.2.1",
    "langchain>=0.3.6",
    "logging>=0.4.9.6",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "wikipedia>=1.4.0",
]
